requirements.txt
pylint
black
numpy
mss
opencv-python
Pillow
